layers.inet.fragment: use fragsize=8 if argument is lower than that
fragment was raising a ZeroDivisionError when given a frag size < 8.
This is a regression, as versions older than 2.4.4 didn't have this
problem. It was a side-effect of fixing fragment to support frag sizes
that are not multiples of 8.

This commit simply tops up frag size to 8 if it's lower, as IP can't
really handle fragments smaller than that anyway, due to the way the
Fragment Offset field works.