[LLMNR] recognize the "T"entative bit
According to https://datatracker.ietf.org/doc/html/rfc4795#section-2.1.1
the TC bit is followed by the T bit:
```
                                      1  1  1  1  1  1
        0  1  2  3  4  5  6  7  8  9  0  1  2  3  4  5
      +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
      |                      ID                       |
      +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
      |QR|   Opcode  | C|TC| T| Z| Z| Z| Z|   RCODE   |
      +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
```
where the 'T'entative bit is set in a response if the
responder is authoritative for the name, but has not yet
verified the uniqueness of the name.

This patch splits the 2-bit "TC" field into two 1-bit fields
corresponding to the "TC" and "T" bits.

The test is added to make sure that tentative responses can be
dissected. It was also cross-checked with Wireshark:
```
Link-local Multicast Name Resolution (response)
    Transaction ID: 0x87df
    Flags: 0x8100 Standard query response, No error
        1... .... .... .... = Response: Message is a response
        .000 0... .... .... = Opcode: Standard query (0)
        .... .0.. .... .... = Conflict: The name is considered unique
        .... ..0. .... .... = Truncated: Message is not truncated
        .... ...1 .... .... = Tentative: Tentative
        .... .... .... 0000 = Reply code: No error (0)
```