Resolve Issue_4006
Sendpfast accepts a float for the PPS value as well as the underlying tcpreplay; however, this is converted to an integer in the tcpreplay string causing flows needing refined throughput values to be inaccurate.
This is a very simple fix by changing it from an integer conversion to a float. I have been doing this manually using sed in docker build to patch this, but should have submitted this when originally found.