CI: run scapy on aarch64, i386, ppc64le, s390x and x86_64 on PRs
The script takes the Fedora package, edits the spec file to make it
compatible with the upstream test suite and then it's all run on all
those architectures on the latest stable Fedora release and Fedora
Rawhide. (Rawhide is kind of a testing relase but it's useful in terms of
catching things like https://github.com/secdev/scapy/issues/4280
reproducible with relatively new packages only).

It was originally prompted by
https://github.com/secdev/scapy/issues/3847 (where the Debian
autopkgtest was run on big-endian and 32-bit machines) and should
hopefully make it easier to catch various issues before they land.

It has been tested since the beginning of 2023 so it should be stable at
this point.