libpcap: make L3socket send() follow routing (#4398)
This makes the L3pcapSocket behave similarly to the other L3*Socket(s).
In a L3 socket, the destination is used to chose the interface to send
the packet on. This means that Scapy should be able to created another
pcap file descriptor if the currently bound one isn't on the proper
interface.