layers/netflow: NFv9/10 flow record type fix
This commit fixes the Python types of a number of fields defined
in Netflowv9/10 for the following pre-defined fields (mostly
the ones in RFC3954 https://datatracker.ietf.org/doc/html/rfc3954).

It appears that in absence of an assigned field type, it defaults
to BytesField instead of IntField. Fields:

* IN_BYTES, IN_PKTS, OUT_BYTES, OUT_PKTS: IntField (4 bytes). Please note
  code doesn't seem to support the larget 8byte optional length.
* MUL_DST_PKTS, MUL_DST_BYTES: same as above
* INPUT_SNMP, OUTPUT_SNMP: ShortField (2 bytes), iface index. Note: code
  doesn't seem to support optional larger values (undefined max. length,
  but 8 bytes might be sufficient).

Note there are _many_ other field types defined that are not part
of the spec / unknown without field type and length.

Fixes #4810

Signed-off-by: Marc Sune <marcdevel@gmail.com>